Clean up lexical parsing XML text nodes
