+++ Cargo.toml
@@ -21,6 +21,7 @@ serde_json = { version = "1.0.79", features = ["unbounded_depth"] }
 
 [dependencies.pyo3]
 version = "0.19.2"
-# https://stackoverflow.com/a/71757467
+# Make the extension-module feature optional and default.
+# See https://pyo3.rs/v0.21.2/faq.html#i-cant-run-cargo-test-or-i-cant-build-in-a-cargo-workspace-im-having-linker-issues-like-symbol-not-found-or-undefined-reference-to-_pyexc_systemerror for more details
 extension-module = ["pyo3/extension-module"]
 default = ["extension-module"]
